Move domain services to infra and drop useless interfaces
