--- src/Nidup/Architool/Domain/SpecFileMover.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Nidup\Architool\Domain;
-
-use Nidup\Architool\Domain\Model\SpecFile;
-
-interface SpecFileMover
-{
-    public function move(SpecFile $specFile);
-}
